requirements.txt
streamlit==1.29.0
pandas==2.1.3
numpy==1.24.3
torch==2.1.0
alpaca-trade-api==3.1.1
newsapi-python==0.2.7
transformers==4.35.0
scikit-learn==1.3.2
plotly==5.18.0
matplotlib==3.8.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0